Move internal/url into pkg/urlt
